requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap4==2022.1
distro==1.9.0
Django==4.1
django-crispy-forms==2.0
django-environ==0.11.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
psycopg2-binary
pydantic==2.10.6
pydantic_core==2.27.2
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
